+++ src/leetcode/122_best_time_to_buy_and_sell_stock_II.cpp
@@ -2,7 +2,7 @@
 
 int maxProfit(std::vector<int>& prices) {
     int profit = 0;
-    for ( int i=0; i < (prices.size() - 1); i++ ) {
+    for ( int i=0; i < ((int)prices.size() - 1); i++ ) {
         int diff = prices[i+1] - prices[i];
         if (diff > 0) {
             profit += diff;
