+++ src/leetcode/14_longest_common_prefix.cpp
@@ -2,10 +2,10 @@
 
 std::string longestCommonPrefix(std::vector<std::string>& strs) {
     if (!strs.size()) return "";
-    for (int i = 0; i < strs[0].size(); i++) {
+    for (int i = 0; i < (int)strs[0].size(); i++) {
         char c = strs[0][i];
-        for (int j = 1; j < strs.size(); j++) {
-            if (i == strs[j].size() || strs[j][i] != c) {
+        for (int j = 1; j < (int)strs.size(); j++) {
+            if (i == (int)strs[j].size() || strs[j][i] != c) {
                 return strs[0].substr(0, i);
             }
         }
