+++ src/leetcode/189_rotate_array.cpp
@@ -12,7 +12,7 @@ void reverse(std::vector<int>& nums, int start, int end) {
 }
     
 void rotate(std::vector<int>& nums, int k) {
-    int size = nums.size();
+    int size = static_cast<int>(nums.size());
     k = k % size;
     reverse(nums, 0, size-1);
     reverse(nums, 0, k-1);
