+++ src/leetcode/26_remove_duplicates_from_sorted_array.cpp
@@ -15,7 +15,7 @@ int removeDuplicates(std::vector<int>& nums) {
         }
     }
     
-    return nums.size();
+    return static_cast<int>(nums.size());
 }
 
 int main() {
