+++ src/leetcode/8_string_to_integer_atoi.cpp
@@ -5,20 +5,20 @@ int myAtoi(std::string s) {
     int multiplier = 1;
     int result = 0;
     int upperIntBound = INT_MAX / 10;
-    if (s.length() == 0) return 0;
+    if ((int)s.length() == 0) return 0;
     
-    while (i < s.length() && s[i] == ' ') {
+    while (i < (int)s.length() && s[i] == ' ') {
         i++;
     }
     
-    if (i < s.length() && (s[i] == '+' || s[i] == '-' )) {
+    if (i < (int)s.length() && (s[i] == '+' || s[i] == '-' )) {
         multiplier = (s[i++] == '-') ? -1 : 1;
     }
     
     int currentVal = 0;
     int maxCurrentVal = (multiplier == 1) ? 7 : 8;
     int diff = 0;
-    while (i < s.length() && s[i] >= '0' && s[i] <= '9') {
+    while (i < (int)s.length() && s[i] >= '0' && s[i] <= '9') {
         currentVal = s[i] - '0';    
         diff = result - upperIntBound;
         if (diff > 0 || (diff == 0 && currentVal >= maxCurrentVal)) {
